Fixed issue with items not working outside of the game
- Some other changes

+++ src/main/java/hu/montlikadani/RageMode/signs/SignData.java
@@ -23,9 +23,10 @@ import hu.montlikadani.ragemode.utils.MaterialUtil;
 public class SignData {
 
 	private String world;
-	private int x;
-	private int y;
-	private int z;
+
+	private double x;
+	private double y;
+	private double z;
 
 	private String game;
 	private SignPlaceholder placeholder;
@@ -44,15 +45,15 @@ public class SignData {
 		return world;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public void setZ(int z) {
+	public void setZ(double z) {
 		this.z = z;
 	}
